+++ generators/php/sdk/src/endpoint/request/WrappedEndpointRequestGenerator.ts
@@ -8,6 +8,7 @@ import { FernIr } from "@fern-fern/ir-sdk";
 import { DefaultValueExtractor } from "../../DefaultValueExtractor.js";
 import { SdkCustomConfigSchema } from "../../SdkCustomConfig.js";
 import { SdkGeneratorContext } from "../../SdkGeneratorContext.js";
+import { mayOmitRequestBody } from "../utils/mayOmitRequestBody.js";
 
 export declare namespace WrappedEndpointRequestGenerator {
     export interface Args {
@@ -129,12 +130,15 @@ export class WrappedEndpointRequestGenerator extends FileGenerator<
 
         this.endpoint.requestBody?._visit({
             reference: (reference) => {
+                const bodyType = this.context.phpTypeMapper.convert({ reference: reference.requestBodyType });
                 this.addFieldWithMethods({
                     clazz,
                     name: this.wrapper.bodyKey,
                     field: php.field({
                         name: this.context.getPropertyName(this.wrapper.bodyKey),
-                        type: this.context.phpTypeMapper.convert({ reference: reference.requestBodyType }),
+                        type: mayOmitRequestBody({ context: this.context, endpoint: this.endpoint })
+                            ? php.Type.optional(bodyType)
+                            : bodyType,
                         access: this.context.getPropertyAccess(),
                         docs: reference.docs
                     }),