+++ seed/php-sdk/respect-optional-request-body/src/dynamic-snippets/example0/snippet.php
@@ -4,6 +4,7 @@ namespace Example;
 
 use Seed\SeedClient;
 use Seed\Requests\RefundBody;
+use Seed\Types\RefundRequest;
 
 $client = new SeedClient(
     options: [
@@ -12,5 +13,7 @@ $client = new SeedClient(
 );
 $client->refund(
     'refund-id',
-    new RefundBody([]),
+    new RefundBody([
+        'body' => new RefundRequest([]),
+    ]),
 );