+++ seed/swift-sdk/respect-optional-request-body/Tests/Snippets/Example4.swift
@@ -5,11 +5,8 @@ enum Example4 {
     static func snippet() async throws {
         let client = ApiClient(baseURL: "https://api.fern.com")
 
-        _ = try await client.requiredRefund(
-            id: "id",
-            request: RefundRequest(
-                amount: 1.1
-            )
-        )
+        _ = try await client.bulkRefund(request: RefundRequest(
+
+        ))
     }
 }